cc spi module compiling

+++ mrflib/src/iqueue.h
@@ -18,5 +18,5 @@ void queue_init(IQUEUE *q);
 int queue_full(IQUEUE *q);
 int queue_data_avail(IQUEUE *q);
 int queue_push(IQUEUE *q, uint8 data);
-uint8 queue_pop(IQUEUE *q);
+int16 queue_pop(IQUEUE *q);
 #endif
